modified : libft.h

+++ headers/libft.h
@@ -1,6 +1,6 @@
 #ifndef	FT_MEMSET
 #define	FT_MEMSET
 
-	void    *ft_memset(void *str, char chr, size_t n);
+	void    *ft_memset(void *str, char chr, int n);
 
 #endif
